mcp-dns-intel/pyproject.toml
[project]
name = "mcp-dns-intel"
version = "1.0.0"
description = "WHOIS lookup, DNS records, subdomain enumeration, SSL certificate analysis"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "dnspython>=2.6.0",
    "python-whois>=0.9.4",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "cryptography>=42.0.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project.scripts]
mcp-dns-intel = "src.server:main"
